requirements.txt
sentence-transformers
faiss-cpu
numpy
openai
flask
flask-cors
python-dotenv
pypdf
